requirements.txt
plantcv
pillow
numpy